move test directory.
